+++ README.md
@@ -1,1 +1,3 @@
 Backend repository for BlockBudget Project
+
+#Addition of payment modules if needed.
