+++ contracts/BlockBudget.sol
@@ -1,0 +1,124 @@
+// SPDX-License-Identifier: MIT
+
+/**
+ * @title BlockBudget
+ * @notice Main contract integrating all personal finance management features
+ * @dev Serves as the central hub for user interaction and financial management
+ */
+contract BlockBudget {
+    /* ========== USER PROFILE MANAGEMENT ========== */
+    
+    /**
+     * @notice Functions for managing user accounts and profiles
+     * @dev Required parameters for profile management:
+     * - userRegistration: Create new user accounts
+     *   - username
+     *   - email (optional, hashed)
+     *   - walletAddress
+     *   - preferredCurrency
+     *   - notificationPreferences
+     * 
+     * - profileConfiguration:
+     *   - privacySettings: Enum (PUBLIC, PRIVATE, FRIENDS_ONLY)
+     *   - displayPreferences
+     *   - languagePreference
+     *   - timeZone
+     * 
+     * - walletManagement:
+     *   - primaryWallet
+     *   - additionalWallets
+     *   - walletLabels
+     *   - connectionStatus
+     */
+
+    /* ========== BUDGET MANAGEMENT ========== */
+
+    /**
+     * @notice Functions for creating and managing budgets
+     * @dev Include functions for:
+     * - createBudget: Set up new budget plans
+     *   - timeframe
+     *   - categories
+     *   - limits
+     * 
+     * - budgetTracking:
+     *   - trackSpending
+     *   - categoryAllocations
+     *   - limitAdjustments
+     * 
+     * - budgetAnalysis:
+     *   - spendingPatterns
+     *   - categoryBreakdown
+     *   - trendAnalysis
+     *   - forecastGeneration
+     */
+
+    /* ========== EXPENSE TRACKING ========== */
+
+    /**
+     * @notice Functions for expense management
+     * @dev Include functions for:
+     * - recordExpense: Log new expenses
+     *   - amount
+     *   - category
+     *   - date
+     *   - description
+     *   - attachments
+     * 
+     * - expenseCategories:
+     *   - createCategory
+     *   - modifyCategory
+     *   - mergeCategories
+     * 
+     * - recurringExpenses:
+     *   - setupRecurring
+     *   - modifyRecurring
+     *   - cancelRecurring
+     */
+
+    /* ========== INCOME MANAGEMENT ========== */
+
+    /**
+     * @notice Functions for tracking and managing income
+     * @dev Include functions for:
+     * - recordIncome: Log income entries
+     *   - source
+     *   - amount
+     *   - frequency
+     * 
+     * - incomeCategories:
+     *   - salaryIncome
+     *   - investmentIncome
+     *   - passiveIncome
+     * 
+     * - recurringIncome:
+     *   - scheduleRecurring
+     *   - trackPayments
+     *   - forecastIncome
+     */
+
+    /* ========== FINANCIAL GOALS INTEGRATION ========== */
+
+    /**
+     * @notice Functions for managing financial goals
+     * @dev Include functions for:
+     * - linkGoalBasedSavings: Connect to GoalBasedLib
+     * - linkTimeLockedSavings: Connect to TimeLockedLib
+     * - linkContributions: Connect to ContributionLib
+     * - trackGoalProgress: Monitor all savings goals
+     * - aggregatePerformance: Combine goal metrics
+     */
+
+     
+    /* ========== ADMIN FUNCTIONS ========== */
+
+    /**
+     * @notice Administrative functions
+     * @dev Include functions for:
+     * - systemMaintenance: Routine updates
+     * - userManagement: Account administration
+     * - troubleshooting: Issue resolution
+     * - systemUpgrades: Contract improvements
+     * - emergencyControls: Crisis management
+     */
+}
