+++ contracts/ContributionLib.sol
@@ -1,0 +1,138 @@
+// SPDX-License-Identifier: MIT
+
+/**
+ * @title ContributionLib
+ * @notice Library for managing group contributions and savings pools
+ * @dev Handles creation and management of contribution groups
+ */
+contract ContributionLib {
+    /* ========== CONTRIBUTION GROUP CREATION ========== */
+    
+    /**
+     * @notice Creates a new contribution group with specified parameters
+     * @dev Required parameters for group creation:
+     * - groupName: Name of the contribution group
+     * - contributionType: Enum (FAMILY, ORGANIZATION, PUBLIC)
+     * - minContribution: Minimum amount required per contribution
+     * - maxMembers: Maximum number of allowed participants
+     * - frequency: Enum (DAILY, WEEKLY, MONTHLY)
+     * - frequencyCount: Number of contribution cycles
+     * - privacyType: Enum (PUBLIC, PRIVATE)
+     * - distributionMethod: Enum (ROUND_ROBIN, NEEDS_BASED, MILESTONE_BASED)
+     * - isMultisigRequired: Boolean for multisig requirement
+     * - multisigApprovers: Array of addresses if multisig is required
+     * - minApprovalsRequired: Number of required approvals for multisig
+     */
+
+    /* ========== MEMBERSHIP MANAGEMENT ========== */
+
+    /**
+     * @notice Functions for managing group membership
+     * @dev Include functions for:
+     * - joinGroup: Allow users to join (requires validation based on privacy type)
+     * - leaveGroup: Process member exit (with rules for fund handling)
+     * - inviteMember: Admin function to invite specific addresses (for private groups)
+     * - kickMember: Admin function to remove problematic members
+     * - transferMembership: Allow membership transfer under specific conditions
+     */
+
+    /* ========== CONTRIBUTION OPERATIONS ========== */
+
+    /**
+     * @notice Functions handling contribution processes
+     * @dev Include functions for:
+     * - makeContribution: Process regular contributions
+     * - validateContribution: Check if contribution meets requirements
+     * - recordContribution: Track contribution history
+     * - getContributionStatus: View member's contribution status
+     * - handleLateFees: Process penalties for late contributions
+     * - exemptContribution: Admin function to excuse missed contributions
+     */
+
+    /* ========== DISTRIBUTION MANAGEMENT ========== */
+
+    /**
+     * @notice Functions for managing fund distribution
+     * @dev Implementation based on distribution method:
+     * 
+     * For ROUND_ROBIN:
+     * - initiateRoundRobinCycle: Start new distribution cycle
+     * - determineNextRecipient: Select next member to receive funds
+     * - processRoundRobinPayment: Execute payment to current recipient
+     * 
+     * For NEEDS_BASED:
+     * - submitNeedRequest: Member submits funding request
+     * - voteOnRequest: Members vote on funding requests
+     * - processNeedsBasedPayment: Execute approved payments
+     * 
+     * For MILESTONE_BASED:
+     * - defineMilestone: Set milestone criteria
+     * - validateMilestone: Check if milestone is achieved
+     * - processMilestonePayment: Distribute funds upon milestone completion
+     */
+
+    /* ========== MULTISIG FUNCTIONALITY ========== */
+
+    /**
+     * @notice Functions for multisig operations
+     * @dev Include functions for:
+     * - proposeTransaction: Create new transaction proposal
+     * - approveTransaction: Record approver's confirmation
+     * - executeTransaction: Process transaction after sufficient approvals
+     * - revokeApproval: Allow approvers to change their vote
+     * - changeApprovers: Modify list of approved signers
+     * - updateMinApprovals: Change required number of approvals
+     */
+
+    /* ========== ADMINISTRATIVE FUNCTIONS ========== */
+
+    /**
+     * @notice Functions for group administration
+     * @dev Include functions for:
+     * - modifyGroupRules: Update group parameters
+     * - pauseGroup: Temporarily halt operations
+     * - dissolveGroup: End group and distribute remaining funds
+     * - generateReport: Create contribution/distribution reports
+     * - handleDispute: Process and resolve member disputes
+     * - updatePrivacySettings: Modify group privacy settings
+     */
+
+    /* ========== EMERGENCY OPERATIONS ========== */
+
+    /**
+     * @notice Functions for handling emergencies
+     * @dev Include functions for:
+     * - initiateEmergencyWithdrawal: Process urgent withdrawal requests
+     * - freezeOperations: Halt all transactions in emergency
+     * - unfreezeOperations: Resume normal operations
+     * - emergencyExit: Allow members to exit during crisis
+     */
+
+    /* ========== VIEW FUNCTIONS ========== */
+
+    /**
+     * @notice Read-only functions for accessing group data
+     * @dev Include functions for:
+     * - getGroupDetails: Retrieve group configuration
+     * - getMemberList: List all current members
+     * - getContributionHistory: View historical contributions
+     * - getDistributionSchedule: View upcoming distributions
+     * - getGroupMetrics: Calculate group performance metrics
+     * - getMultisigStatus: Check status of multisig proposals
+     */
+
+    /* ========== EVENTS ========== */
+
+    /**
+     * @notice Events to be emitted for important operations
+     * @dev Should include events for:
+     * - GroupCreated: New group creation
+     * - MembershipChanged: Member join/leave
+     * - ContributionReceived: New contribution
+     * - DistributionProcessed: Funds distribution
+     * - RulesModified: Group parameter updates
+     * - EmergencyActionTriggered: Emergency operations
+     * - MultisigProposalCreated: New multisig proposal
+     * - MultisigApprovalChanged: Approval status change
+     */
+}
