+++ contracts/FactoryContract.sol
@@ -1,0 +1,30 @@
+/**
+ * @title FinanceFactory
+ * @notice Factory contract for deploying personal finance components
+ * @dev Creates and manages instances of all financial contracts
+ */
+contract BlockBudgetFactory {
+    /**
+     * @notice Deployment functions
+     * @dev Creates new instances of:
+     * - Contribution groups
+     * - Time-locked savings
+     * - Goal-based savings
+     */
+
+    /**
+     * @notice Contract management
+     * @dev Handles:
+     * - Contract upgrades
+     * - Security checks
+     * - Instance tracking
+     */
+
+    /**
+     * @notice Access control
+     * @dev Implements:
+     * - Role management
+     * - Permission settings
+     * - Admin functions
+     */
+}
