+++ contracts/GoalBasedLib.sol
@@ -1,0 +1,132 @@
+// SPDX-License-Identifier: MIT
+
+/**
+ * @title GoalBasedLib
+ * @notice Library for managing goal-based savings targets and milestones
+ * @dev Implements functionality for personal and group savings goals
+ */
+contract GoalBasedLib {
+    /* ========== GOAL CREATION AND SETUP ========== */
+
+    /**
+     * @notice Functions for creating and initializing savings goals
+     * @dev Required parameters for goal creation:
+     * - goalName: Name/description of the savings goal
+     * - targetAmount: Final amount to be saved
+     * - deadline: Timeline for achieving the goal
+     * - goalType: Enum (PERSONAL, GROUP, EDUCATIONAL, PURCHASE, INVESTMENT)
+     * - savingFrequency: Enum (DAILY, WEEKLY, MONTHLY)
+     * - minContributionAmount: Minimum amount per contribution
+     * - isFlexible: Boolean for flexible vs. fixed contributions
+     * - autoContribute: Boolean for automatic contributions
+     * - milestones: Array of milestone percentages and rewards
+     * - penaltyRate: Early withdrawal penalty percentage
+     */
+
+    /* ========== MILESTONE MANAGEMENT ========== */
+
+    /**
+     * @notice Functions for managing savings milestones
+     * @dev Include functions for:
+     * - defineMilestone: Create new milestone targets
+     * - updateMilestone: Modify existing milestone parameters
+     * - trackMilestoneProgress: Monitor progress towards milestones
+     * - validateMilestone: Verify milestone achievement
+     * - distributeMilestoneRewards: Process reward distribution
+     * - getMilestoneHistory: View completed milestones
+     */
+
+    /* ========== PROGRESS TRACKING ========== */
+
+    /**
+     * @notice Functions for monitoring savings progress
+     * @dev Include functions for:
+     * - calculateProgress: Compute percentage completion
+     * - generateProgressReport: Create detailed progress summary
+     * - projectCompletion: Estimate completion date
+     * - checkMilestoneStatus: Verify milestone achievements
+     * - trackSavingRate: Monitor saving speed
+     * - compareToSchedule: Check if on track with timeline
+     */
+
+    /* ========== REWARD SYSTEM ========== */  //if needed
+
+    /**
+     * @notice Functions for managing achievement rewards
+     * @dev Include functions for:
+     * - defineRewards: Set up reward structure
+     * - calculateReward: Compute earned rewards
+     * - distributeRewards: Process reward payments
+     * - modifyRewardRules: Update reward parameters
+     * - checkRewardEligibility: Verify reward qualification
+     * - getRewardHistory: View distributed rewards
+     */
+
+    /* ========== GOAL COMPLETION ========== */
+
+    /**
+     * @notice Functions for handling goal achievement
+     * @dev Include functions for:
+     * - verifyCompletion: Check if goal is achieved
+     * - processCompletion: Handle goal completion
+     * - extendGoal: Modify goal timeline
+     * - increaseTarget: Adjust target amount
+     * - finalizeGoal: Close completed goal
+     * - generateCompletionReport: Summary of achievement
+     */
+
+    /* ========== WITHDRAWAL MANAGEMENT ========== */
+
+    /**
+     * @notice Functions for managing fund withdrawals
+     * @dev Include functions for:
+     * - regularWithdrawal: Process planned withdrawals
+     * - emergencyWithdrawal: Handle urgent withdrawals
+     * - calculatePenalty: Compute early withdrawal fees
+     * - approveWithdrawal: Validate withdrawal request
+     * - processWithdrawal: Execute fund transfer
+     * - getWithdrawalHistory: View past withdrawals
+     */
+
+    /* ========== EMERGENCY HANDLING ========== */
+
+    /**
+     * @notice Functions for emergency situations
+     * @dev Include functions for:
+     * - modifyGoal: Update goal parameters
+     * - freezeGoal: Temporarily suspend goal
+     * - cancelGoal: Terminate goal early
+     * - emergencyWithdraw: Process urgent withdrawals
+     * - updatePenalties: Modify penalty structure
+     * - handleDisputes: Resolve issues
+     */
+
+    /* ========== VIEW FUNCTIONS ========== */
+
+    /**
+     * @notice Read-only functions for accessing goal data
+     * @dev Include functions for:
+     * - getGoalDetails: Retrieve goal configuration
+     * - getProgress: View current progress
+     * - getMilestones: List all milestones
+     * - getContributionHistory: View past contributions
+     * - calculateProjections: Estimate future progress
+     * - getRewardStatus: Check available rewards
+     */
+
+    /* ========== EVENTS ========== */
+
+    /**
+     * @notice Events to be emitted for important operations
+     * @dev Should include events for:
+     * - GoalCreated: New goal initiation
+     * - ContributionMade: New contribution received
+     * - MilestoneAchieved: Milestone completion
+     * - RewardDistributed: Reward payment
+     * - GoalModified: Parameter updates
+     * - WithdrawalProcessed: Fund withdrawal
+     * - GoalCompleted: Target achievement
+     * - EmergencyAction: Emergency operations
+     */
+
+}
