+++ contracts/TimeLockedLib.sol
@@ -1,0 +1,125 @@
+// SPDX-License-Identifier: MIT
+
+/**
+ * @title TimeLockedLib
+ * @notice Library for managing time-locked savings with interest mechanisms
+ * @dev Implements time-based restrictions and interest calculations
+ */
+contract TimeLockedLib {
+    /* ========== ACCOUNT CREATION AND SETUP ========== */
+    
+    /**
+     * @notice Functions for creating and initializing time-locked accounts
+     * @dev Required parameters for account creation:
+     * - accountType: Enum (FIXED_TERM, FLEXIBLE_TERM, LADDER_TERM)
+     * - lockDuration: Time period for fund lock
+     * - unlockDate: Specific date for fund release
+     * - interestType: Enum (FIXED, VARIABLE, COMPOUND)
+     * - interestRate: Base interest rate percentage
+     * - minDepositAmount: Minimum required deposit
+     * - maxDepositAmount: Maximum allowed deposit
+     * - earlyWithdrawalPenalty: Penalty percentage
+     * - allowPartialWithdrawal: Boolean for partial withdrawals
+     * - allowAdditionalDeposits: Boolean for extra deposits
+     */
+
+    /* ========== DEPOSIT MANAGEMENT ========== */
+
+    /**
+     * @notice Functions for handling deposits
+     * @dev Include functions for:
+     * - initialDeposit: Process first-time deposits
+     * - additionalDeposit: Handle subsequent deposits
+     * - validateDeposit: Check deposit constraints
+     * - extendLockPeriod: Increase lock duration
+     * - upgradeAccount: Modify account parameters
+     * - getDepositHistory: View deposit records
+     * - calculateTotalDeposits: Sum all deposits
+     */
+
+    /* ========== INTEREST CALCULATIONS ========== */
+
+    /**
+     * @notice Functions for managing interest
+     * @dev Include functions for:
+     * - calculateBaseInterest: Compute basic interest
+     * - calculateCompoundInterest: Handle compound interest
+     * - updateInterestRate: Modify current rates
+     * - distributeInterest: Process interest payments
+     * - projectInterestEarnings: Estimate future earnings
+     * - getInterestHistory: View interest payment history
+     * - adjustRatesByDuration: Scale rates by lock time
+     */
+
+    /* ========== WITHDRAWAL PROCESSING ========== */
+
+    /**
+     * @notice Functions for managing withdrawals
+     * @dev Include functions for:
+     * - requestWithdrawal: Initiate withdrawal process
+     * - validateWithdrawalEligibility: Check time constraints
+     * - calculatePenalty: Compute early withdrawal fees
+     * - processPartialWithdrawal: Handle partial withdrawals
+     * - processFullWithdrawal: Process complete withdrawals
+     * - cancelWithdrawalRequest: Reverse withdrawal request
+     * - getWithdrawalHistory: View withdrawal records
+     */
+
+    /* ========== LOCK PERIOD MANAGEMENT ========== */
+
+    /**
+     * @notice Functions for managing time locks
+     * @dev Include functions for:
+     * - checkLockStatus: Verify current lock state
+     * - extendLockPeriod: Increase lock duration
+     * - calculateTimeRemaining: Compute remaining lock time
+     * - modifyUnlockDate: Update unlock timeline
+     * - handleMaturity: Process matured accounts
+     * - autoRenewal: Handle automatic renewals
+     */
+
+    /* ========== INTEREST DISTRIBUTION ========== */
+
+    /**
+     * @notice Functions for interest payment management
+     * @dev Include functions for:
+     * - scheduleInterestPayment: Plan interest distribution
+     * - processInterestPayment: Execute interest transfers
+     * - reinvestInterest: Handle interest reinvestment
+     * - modifyInterestSchedule: Update payment timeline
+     * - calculateAccruedInterest: Compute current earnings
+     * - handleMissedPayments: Process delayed payments
+     */
+     
+
+    /* ========== VIEW FUNCTIONS ========== */
+
+    /**
+     * @notice Read-only functions for accessing account data
+     * @dev Include functions for:
+     * - getAccountDetails: Retrieve account configuration
+     * - getLockStatus: Check current lock state
+     * - getInterestEarned: View accumulated interest
+     * - getMaturityDate: Check account maturity
+     * - getPenaltyRates: View withdrawal penalties
+     * - getProjectedEarnings: Calculate future returns
+     */
+
+    /* ========== EVENTS ========== */
+
+    /**
+     * @notice Events to be emitted for important operations
+     * @dev Should include events for:
+     * - AccountCreated: New account creation
+     * - DepositReceived: New deposit
+     * - InterestPaid: Interest distribution
+     * - WithdrawalProcessed: Withdrawal completion
+     * - LockModified: Lock period changes
+     * - PenaltyCharged: Penalty application
+     * - EmergencyAction: Emergency operations
+     * - AccountClosed: Account termination
+     */
+
+    /* ========== REPORTING AND ANALYTICS ========== */
+
+}
